refactor: Split inspectify backend up into multiple files
